[JENKINS-60579] Move tests on ViewPropertyTest.
